refactor: make service abstract
