+++ src/main/java/subway/service/Service.java
@@ -1,0 +1,13 @@
+package subway.service;
+
+public interface Service {
+
+    void run();
+
+    void add();
+
+    void delete();
+
+    void show();
+
+}
